pyproject.toml
[tool.poetry]
name = "statsservice"
version = "0.4.0"
description = "Stats Service for MONARC."
authors = ["Cédric Bonhomme <cedric@cedricbonhomme.org>"]
license = "AGPL-3.0-or-later"

readme = "README.md"

homepage = "https://github.com/monarc-project/stats-service"
repository = "https://github.com/monarc-project/stats-service"
documentation = "https://www.monarc.lu/documentation/stats-service"

keywords = ["monarc", "security", "stats", "flask", "api"]

classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Web Environment",
    "Topic :: Security",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)"
]

include = [
    "README.md",
    "COPYING",
    "runserver.py"
]

packages = [
    { include = "statsservice" },
    { include = "instance" },
]

[tool.poetry.scripts]
monarc-stats-service = "runserver:run"

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
Flask = "^2.0.2"
requests = "^2.25.1"
flask_restx = "^0.5.1"
jsonschema = "^3.2.0"
Flask-SQLAlchemy = "^2.5.1"
psycopg2-binary = "^2.8.6"
flask_principal = "^0.4.0"
flask_login = "^0.5.0"
Flask-Babel = "^2.0.0"
alembic = "^1.5.8"
Flask-Migrate = "^2.7.0"
pymosp = "^0.4.2"
packaging = "^20.9"
python-dateutil = "^2.8.2"
mypy = "^0.931"
typing-extensions = "^4.0.1"
types-setuptools = "^57.4.7"
types-requests = "^2.27.5"
types-python-dateutil = "^2.8.6"

[tool.poetry.dev-dependencies]
sphinx = "^3.0.3"
sphinx-book-theme = "^0.1.0"
flake8 = "^3.9.1"
pytest = "^6.2.3"
gitchangelog = "^3.0.4"
sphinx-multiversion = "^0.2.4"
pre-commit = "^2.18.1"
black = "^22.3.0"
pyupgrade = "^2.32.0"
reorder-python-imports = "^3.0.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.mypy]
python_version = "3.10"
check_untyped_defs = true
ignore_errors = false
ignore_missing_imports = true
strict_optional = true
no_implicit_optional = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
warn_unreachable = true

show_error_context = true
pretty = true

exclude = "migrations|docs|statsservice/commands"
